Added in a SpriteKit test scene
Got some square bubbles automatically appearing behind the normal UIKit
stuff that already existed in the app.

Added the BG as a node and hid it in the storyboard.

Still need to set the scale for the BG to be proper…will probably need
to be separate sprites that animate for the waves at the top.